Fix issue #15 & #17 and move the copyToClipboard function to a seperate function for easier reusability
